refactor(differential): extract shared Heatmap renderer
